Derive media from sdp in Firefox on answer
In Firefox, incoming calls were dropped because of the "User Denied Media
Access" error. This error in turn originates from getUserMedia that responds
with NO_DEVICES_FOUND, which is caused by the fact that audio and video
constraints of the call were set to false.

When .answer() is called in Chrome,
peerConnection.iceConnectionState = checking
peerConnection.iceGatheringState = gathering
peerConnection.signalingState = stable
and getRemoteStreams() returns a non-empty array.

In Firefox, on the other hand,
peerConnection.iceConnectionState = new
peerConnection.iceGatheringState = gathering
peerConnection.signalingState = have-remote-offer
and getRemoteStreams() returns an empty array.

Some experimenting shows that the remote streams are only added by Firefox
after creating an answer to the offer. Unfortunately, this event happens
after getUserMedia is called, so it's useless for determining whether or
not to activate/de-activate audio/video.

To fix the problem, I'm extracting information about the (non-)existence
of the video/audio stream from peerConnection.remoteDescription.
m=audio and m=video implies that the remote peer has supplied an audio
video stream respectively.