Add before_reregister option.
Set this to a function with a single callback argument.
Before JsSIP reregisters it will call the function.
The reregister will happen once you call the callback.
This gives you a chance to change some settings which
will affect the registration (e.g. generate a time-limited
token to register with).