Implements rtpTimeout and byeReceived RTCSession events
If these events have handlers registered, call termination is
postponed. For both events, the event handler receives an event object
that has a 'finish' callback function. The callback function takes a
single argument 'extraHeaders'.

For the rtpTimeout event, the SIP 408 terminate message will be sent
with whatever extra headers are passed to the callback function.

For the byeReceived event, the SIP 200 response to the BYE message
will be sent with whatever extra headers are passed to the callback
function.

We are using these new callbacks so that we can send call
quality/statistics data via extra SIP headers.