Fixes broken quoted_string_clean and display_name parser rules
This fixes 2 problems with the quoted_string_clean rule:
1. The substring operation (`input.substring(pos-1, offset+1);`) to
remove outer quotes returns incorrect results for strings that
begin with SWS
2. It does not completely unescape the string. The outer DQUOTES are
removed but quoted_pair sequences are not unescaped.

The display_name rule is changed in the following ways:
1. quoted_string is changed to quoted_string_clean so that
display_name is properly unescaped when it matches the quoted_string
form of the rule. trimming and unescaping is no longer the
responsibility of display_name
2. logic has been added to collapse arrays of tokens into a string
when the rule matches the `token ( LWS token )*` form

It is worth mentioning that LWS whitespace occuring in the token form
of display_name is now collapsed into a single SP where as before
it was left intact. This is compliant with the spec. (RFC3261 25)

Additionally, a link to the RFC errata documenting the display_name
rule correction has been included to supplement the 'previous rule is
corrected from RFC3261' comment.